+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/Translation.hs
@@ -21,7 +21,6 @@ import Cardano.Ledger.Shelley.API (
   CertState (..),
   DState (..),
   EpochState (..),
-  VState (..),
   LedgerState (..),
   NewEpochState (..),
   PState (..),
@@ -32,6 +31,7 @@ import Cardano.Ledger.Shelley.API (
   UTxO (..),
   UTxOState (..),
   Update,
+  VState (..),
  )
 import qualified Cardano.Ledger.Shelley.LedgerState as LS (
   returnRedeemAddrsToReserves,