+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Cert.hs
@@ -24,7 +24,7 @@ import Cardano.Ledger.BaseTypes (ShelleyBase)
 import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.Conway.Core
-import Cardano.Ledger.Conway.Era (ConwayCERT, ConwayDELEG, ConwayPOOL, ConwayVDEL)
+import Cardano.Ledger.Conway.Era (ConwayCERT, ConwayDELEG, ConwayGOVCERT, ConwayPOOL)
 import Cardano.Ledger.Conway.Rules.Deleg (ConwayDelegPredFailure (..))
 import Cardano.Ledger.Conway.Rules.GovCert (ConwayGovCertPredFailure)
 import Cardano.Ledger.Conway.TxCert (ConwayCommitteeCert, ConwayDelegCert, ConwayTxCert (..))
@@ -33,9 +33,9 @@ import Cardano.Ledger.Shelley.API (
   DState,
   DelegEnv (DelegEnv),
   DelplEnv (DelplEnv),
-  VState,
   PState,
   PoolEnv (PoolEnv),
+  VState,
  )
 import Cardano.Ledger.Shelley.Rules (ShelleyPoolPredFailure)
 import Control.DeepSeq (NFData)
@@ -56,57 +56,57 @@ import NoThunks.Class (NoThunks)
 data ConwayCertPredFailure era
   = DelegFailure (PredicateFailure (EraRule "DELEG" era))
   | PoolFailure (PredicateFailure (EraRule "POOL" era))
-  | GovCertFailure (PredicateFailure (EraRule "VDEL" era))
+  | GovCertFailure (PredicateFailure (EraRule "GOVCERT" era))
   deriving (Generic)
 
 deriving stock instance
   ( Show (PredicateFailure (EraRule "DELEG" era))
   , Show (PredicateFailure (EraRule "POOL" era))
-  , Show (PredicateFailure (EraRule "VDEL" era))
+  , Show (PredicateFailure (EraRule "GOVCERT" era))
   ) =>
   Show (ConwayCertPredFailure era)
 
 deriving stock instance
   ( Eq (PredicateFailure (EraRule "DELEG" era))
   , Eq (PredicateFailure (EraRule "POOL" era))
-  , Eq (PredicateFailure (EraRule "VDEL" era))
+  , Eq (PredicateFailure (EraRule "GOVCERT" era))
   ) =>
   Eq (ConwayCertPredFailure era)
 
 instance
   ( NoThunks (PredicateFailure (EraRule "DELEG" era))
   , NoThunks (PredicateFailure (EraRule "POOL" era))
-  , NoThunks (PredicateFailure (EraRule "VDEL" era))
+  , NoThunks (PredicateFailure (EraRule "GOVCERT" era))
   ) =>
   NoThunks (ConwayCertPredFailure era)
 
 instance
   ( NFData (PredicateFailure (EraRule "DELEG" era))
   , NFData (PredicateFailure (EraRule "POOL" era))
-  , NFData (PredicateFailure (EraRule "VDEL" era))
+  , NFData (PredicateFailure (EraRule "GOVCERT" era))
   ) =>
   NFData (ConwayCertPredFailure era)
 
 data ConwayCertEvent era
   = DelegEvent (Event (ConwayDELEG era))
   | PoolEvent (Event (ConwayPOOL era))
-  | GovCertEvent (Event (ConwayVDEL era))
+  | GovCertEvent (Event (ConwayGOVCERT era))
 
 instance
   forall era.
   ( Era era
   , State (EraRule "DELEG" era) ~ DState era
   , State (EraRule "POOL" era) ~ PState era
-  , State (EraRule "VDEL" era) ~ VState era
+  , State (EraRule "GOVCERT" era) ~ VState era
   , Environment (EraRule "DELEG" era) ~ DelegEnv era
   , Environment (EraRule "POOL" era) ~ PoolEnv era
-  , Environment (EraRule "VDEL" era) ~ PParams era
+  , Environment (EraRule "GOVCERT" era) ~ PParams era
   , Signal (EraRule "DELEG" era) ~ ConwayDelegCert (EraCrypto era)
   , Signal (EraRule "POOL" era) ~ PoolCert (EraCrypto era)
-  , Signal (EraRule "VDEL" era) ~ ConwayCommitteeCert (EraCrypto era)
+  , Signal (EraRule "GOVCERT" era) ~ ConwayCommitteeCert (EraCrypto era)
   , Embed (EraRule "DELEG" era) (ConwayCERT era)
   , Embed (EraRule "POOL" era) (ConwayCERT era)
-  , Embed (EraRule "VDEL" era) (ConwayCERT era)
+  , Embed (EraRule "GOVCERT" era) (ConwayCERT era)
   , TxCert era ~ ConwayTxCert era
   ) =>
   STS (ConwayCERT era)
@@ -124,16 +124,16 @@ certTransition ::
   forall era.
   ( State (EraRule "DELEG" era) ~ DState era
   , State (EraRule "POOL" era) ~ PState era
-  , State (EraRule "VDEL" era) ~ VState era
+  , State (EraRule "GOVCERT" era) ~ VState era
   , Environment (EraRule "DELEG" era) ~ DelegEnv era
   , Environment (EraRule "POOL" era) ~ PoolEnv era
-  , Environment (EraRule "VDEL" era) ~ PParams era
+  , Environment (EraRule "GOVCERT" era) ~ PParams era
   , Signal (EraRule "DELEG" era) ~ ConwayDelegCert (EraCrypto era)
   , Signal (EraRule "POOL" era) ~ PoolCert (EraCrypto era)
-  , Signal (EraRule "VDEL" era) ~ ConwayCommitteeCert (EraCrypto era)
+  , Signal (EraRule "GOVCERT" era) ~ ConwayCommitteeCert (EraCrypto era)
   , Embed (EraRule "DELEG" era) (ConwayCERT era)
   , Embed (EraRule "POOL" era) (ConwayCERT era)
-  , Embed (EraRule "VDEL" era) (ConwayCERT era)
+  , Embed (EraRule "GOVCERT" era) (ConwayCERT era)
   , TxCert era ~ ConwayTxCert era
   ) =>
   TransitionRule (ConwayCERT era)
@@ -147,8 +147,8 @@ certTransition = do
     ConwayTxCertPool poolCert -> do
       newPState <- trans @(EraRule "POOL" era) $ TRC (PoolEnv slot pp, certPState, poolCert)
       pure $ cState {certPState = newPState}
-    ConwayTxCertCommittee vDelCert -> do
-      newVState <- trans @(EraRule "VDEL" era) $ TRC (pp, certVState, vDelCert)
+    ConwayTxCertCommittee govCert -> do
+      newVState <- trans @(EraRule "GOVCERT" era) $ TRC (pp, certVState, govCert)
       pure $ cState {certVState = newVState}
 
 instance
@@ -175,10 +175,10 @@ instance
 
 instance
   ( Era era
-  , STS (ConwayVDEL era)
-  , PredicateFailure (EraRule "VDEL" era) ~ ConwayGovCertPredFailure era
+  , STS (ConwayGOVCERT era)
+  , PredicateFailure (EraRule "GOVCERT" era) ~ ConwayGovCertPredFailure era
   ) =>
-  Embed (ConwayVDEL era) (ConwayCERT era)
+  Embed (ConwayGOVCERT era) (ConwayCERT era)
   where
   wrapFailed = GovCertFailure
   wrapEvent = GovCertEvent
@@ -187,7 +187,7 @@ instance
   ( Typeable era
   , EncCBOR (PredicateFailure (EraRule "DELEG" era))
   , EncCBOR (PredicateFailure (EraRule "POOL" era))
-  , EncCBOR (PredicateFailure (EraRule "VDEL" era))
+  , EncCBOR (PredicateFailure (EraRule "GOVCERT" era))
   ) =>
   EncCBOR (ConwayCertPredFailure era)
   where
@@ -201,7 +201,7 @@ instance
   ( Typeable era
   , DecCBOR (PredicateFailure (EraRule "DELEG" era))
   , DecCBOR (PredicateFailure (EraRule "POOL" era))
-  , DecCBOR (PredicateFailure (EraRule "VDEL" era))
+  , DecCBOR (PredicateFailure (EraRule "GOVCERT" era))
   ) =>
   DecCBOR (ConwayCertPredFailure era)
   where